MedNAIS2_extracted/MedNAIS/mednais-analysis/mednais-sop-marketplace-main/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
httpx>=0.28.1
prisma==0.15.0
python-dotenv>=1.0.0
emergentintegrations==0.1.0
PyJWT>=2.8.0
